requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
inflection==0.5.1
influxdb==5.3.1
more-itertools==8.13.0
msgpack==1.0.4
Nasdaq-Data-Link==1.0.2
numpy==1.22.4
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
urllib3==1.26.9
